move comparePositonVertices() to test-helper
