+++ src/api/primitives2d.test.js
@@ -1,23 +1,6 @@
 const test = require('ava')
 const { square, circle, triangle, polygon } = require('./primitives2d-api')
-const { sideEquals, shape2dToNestedArray } = require('./test-helpers')
-
-/* FIXME : not entirely sure how to deal with this, but for now relies on inspecting
-output data structures: we should have higher level primitives ... */
-
-// helper functions
-function comparePositonVertices (obs, exp) {
-  for (let index = 0; index < obs.length; index++) {
-    let side = obs[index]
-    const same = side.vertex0.pos._x === exp[index][0][0] && side.vertex0.pos._y === exp[index][0][1]
-      && side.vertex1.pos._x === exp[index][1][0] && side.vertex1.pos._y === exp[index][1][1]
-    // console.log('side', side.vertex0.pos, same)
-    if (!same) {
-      return false
-    }
-  }
-  return true
-}
+const { sideEquals, shape2dToNestedArray, comparePositonVertices } = require('./test-helpers')
 
 test.failing('triangle (defaults)', t => {
   const obs = triangle()
